requirements.txt
flask
torch
transformers
Pillow
matplotlib
requests
openai
python-dotenv